refactor: Profile domain Member와 분리 (#251)
