+++ src/main/java/com/api/ttoklip/domain/member/service/MemberService.java
@@ -5,7 +5,7 @@ import static com.api.ttoklip.global.exception.ErrorType._USER_NOT_FOUND_DB;
 
 import com.api.ttoklip.domain.member.domain.Member;
 import com.api.ttoklip.domain.member.dto.response.MemberStreetResponse;
-import com.api.ttoklip.domain.member.dto.response.TargetMemberProfile;
+import com.api.ttoklip.domain.profile.TargetMemberProfile;
 import com.api.ttoklip.domain.member.repository.MemberRepository;
 import com.api.ttoklip.domain.privacy.domain.Profile;
 import com.api.ttoklip.domain.privacy.dto.InterestResponse;
