+++ src/main/java/com/api/ttoklip/domain/profile/ProfileLikeRepository.java
@@ -1,6 +1,5 @@
-package com.api.ttoklip.domain.member.repository;
+package com.api.ttoklip.domain.profile;
 
-import com.api.ttoklip.domain.member.domain.ProfileLike;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface ProfileLikeRepository extends JpaRepository<ProfileLike, Long>, ProfileLikeRepositoryCustom {
